refactor: restructure of auth commands
