+++ internal/registry/auth/oauth/doc.go
