+++ internal/registry/auth/oauth/oauth.go
